Fix. High FindTop10EventsAsync сортирует целочисленным делением, топ-10 не является топом

+++ YAP_middle-csharp_Events/YAP_middle-csharp_Events.Infrastructure/Repository/EventRepository.cs
@@ -78,7 +78,7 @@ namespace YAP_middle_csharp_Events.Infrastructure.Repository
             return await _context.Events
                 .AsNoTracking()
                 .Where(x => x.TotalSeats > 0)
-                .OrderByDescending(x => (x.TotalSeats - x.AvailableSeats) / x.TotalSeats)
+                .OrderByDescending(x => (double)(x.TotalSeats - x.AvailableSeats) / x.TotalSeats)
                 .Take(10)
                 .ToListAsync(cancellationToken);
         }
